test(canon): keep the source-length gate green for the spread work
The gate rejects growth in files already over 350 lines. Move the pinia
spread-helper test into the options-API spread module it belongs to, share
one ComponentUsage base literal across the croquis-cf helpers, and lift the
duplicated closure-scope prop-check loop into a helper.
