+++ crates/vize_canon/src/virtual_ts/scope/component_props.rs
@@ -290,22 +290,7 @@ fn generate_closure_component_props_recursive(
             }
 
             // Emit component prop checks for this scope
-            if let Some(usages) = ctx.components_by_scope.get(&scope_id) {
-                for &(idx, usage) in usages {
-                    profile!(
-                        "canon.virtual_ts.component_prop_checks",
-                        generate_component_prop_checks(
-                            ts,
-                            mappings,
-                            usage,
-                            idx,
-                            ctx.template_prop_names,
-                            ctx.source_context,
-                            &vfor_inner_indent,
-                        )
-                    );
-                }
-            }
+            emit_scope_component_prop_checks(ts, mappings, ctx, scope_id, &vfor_inner_indent);
 
             // Recursively handle child closure scopes (v-for and v-slot)
             recurse_child_closure_scopes(ts, mappings, ctx, scope_id, &vfor_inner_indent);
@@ -347,22 +332,7 @@ fn generate_closure_component_props_recursive(
                 }
             }
             // Emit component prop checks for this scope
-            if let Some(usages) = ctx.components_by_scope.get(&scope_id) {
-                for &(idx, usage) in usages {
-                    profile!(
-                        "canon.virtual_ts.component_prop_checks",
-                        generate_component_prop_checks(
-                            ts,
-                            mappings,
-                            usage,
-                            idx,
-                            ctx.template_prop_names,
-                            ctx.source_context,
-                            &inner_indent,
-                        )
-                    );
-                }
-            }
+            emit_scope_component_prop_checks(ts, mappings, ctx, scope_id, &inner_indent);
 
             // Recursively handle child closure scopes (v-for and v-slot)
             recurse_child_closure_scopes(ts, mappings, ctx, scope_id, &inner_indent);
@@ -371,6 +341,33 @@ fn generate_closure_component_props_recursive(
             ts.push_str("};\n");
         }
         _ => {}
+    }
+}
+
+/// Emit the prop checks for every component usage bound to a closure scope.
+fn emit_scope_component_prop_checks(
+    ts: &mut String,
+    mappings: &mut Vec<VizeMapping>,
+    ctx: &VForPropsContext<'_>,
+    scope_id: u32,
+    indent: &str,
+) {
+    let Some(usages) = ctx.components_by_scope.get(&scope_id) else {
+        return;
+    };
+    for &(idx, usage) in usages {
+        profile!(
+            "canon.virtual_ts.component_prop_checks",
+            generate_component_prop_checks(
+                ts,
+                mappings,
+                usage,
+                idx,
+                ctx.template_prop_names,
+                ctx.source_context,
+                indent,
+            )
+        );
     }
 }
